fix: support Minecraft on FAT32 USB storage
Handle USB cache capabilities and unsupported synchronization commands without masking real I/O failures. Retain the first disk error and expose disk capacity, mount state and diagnostics through disks.bin.

Support large FAT32 volumes and high cluster addresses, package a relocatable Minecraft client with corrected Java interpreters, and fix the i386 socket timeout ABI.

Validated with x86_64 and i386 USB regressions and flush-error injection, host ASan/UBSan storage tests, and user-confirmed hardware operation.

+++ kernel/include/define.h
@@ -255,9 +255,9 @@ struct FAT_CACHE {
   int FatMaxTerms;
   unsigned int ClustnoBytes;
   unsigned int RootMaxFiles;
-  unsigned int RootDictAddress;
+  uint64_t RootDictAddress;
   unsigned int FileDataAddress;
-  unsigned int imgTotalSize;
+  uint64_t imgTotalSize;
   unsigned short SectorBytes;
   unsigned int Fat1Address, Fat2Address;
   unsigned char *FatClustnoFlags;
